refactor: move demo files and run ESLint correctly
