+++ package.json
@@ -14,10 +14,10 @@
     "build": "npm run tsc-esm && cheminfo-build --entry lib-esm/index.js --root TreeSVG",
     "dev": "vite",
     "clean": "rimraf lib lib-esm",
-    "prettier": "prettier --check src",
-    "prettier-write": "prettier --write src",
+    "prettier": "prettier --check src demo",
+    "prettier-write": "prettier --write src demo",
     "preview": "vite preview",
-    "eslint": "eslint",
+    "eslint": "eslint src demo",
     "eslint-fix": "npm run eslint -- --fix",
     "prepack": "npm run tsc",
     "tsc": "npm run clean && npm run tsc-cjs && npm run tsc-esm",
