+++ src/components/Arrow.tsx
@@ -1,7 +1,7 @@
 import { Point } from '../types/Point';
 
 import { CenteredText } from './CenteredText';
-import { MarkerDef } from './MarkerDef';
+import { refX } from './MarkerDef.utils';
 import { RightText } from './RightText';
 /**
  * Creates a simple arrow between 2 points. The arrow has one inflection point, which is the middle of the line.
@@ -18,7 +18,6 @@ export function Arrow(props: {
 }) {
   const { from, to, label, labelPosition = 'center' } = props;
   const middle = { x: (from.x + to.x) / 2, y: (from.y + to.y) / 2 };
-  const refX = MarkerDef(props).props.children.props.refX;
 
   const headInflectionPoint = {
     x: to.x - refX * 2, // refX attribute of the marker element in the Arrow component shifts the arrow head to the left, and that the vertices of the arrow head are set to refX, which means that the beginning of the arrow head is refX * 2 to the left.
@@ -26,7 +25,6 @@ export function Arrow(props: {
   };
   const d = `M ${from.x} ${from.y} C ${middle.x} ${from.y}  ${middle.x} ${to.y}  ${headInflectionPoint.x} ${to.y} Q ${headInflectionPoint.x} ${to.y}  ${to.x} ${to.y} `;
 
-  //const d = `M ${from.x} ${from.y} C ${middle.x} ${from.y}  ${middle.x} ${from.y}  ${middle.x} ${middle.y} ${middle.x} ${to.y} ${middle.x} ${to.y}  ${to.x} ${to.y}`;
   return (
     <g>
       <path
