+++ src/components/CenteredText.tsx
@@ -1,6 +1,6 @@
 export function CenteredText(props: { x: number; y: number; label: string }) {
   const { x, y, label } = props;
-  const labelPosition = 'translate(' + x + ',' + y + ')';
+  const labelPosition = `translate(${x},${y})`;
   return (
     <>
       <text
