+++ src/components/MarkerDef.tsx
@@ -1,15 +1,17 @@
-export function MarkerDef(props) {
+import { refX } from './MarkerDef.utils';
+
+export function MarkerDef() {
   return (
     <defs>
       <marker
         id="arrowhead"
         markerWidth="10"
         markerHeight="7"
-        refX="10"
+        refX={refX}
         refY="3.5"
         orient="auto"
       >
-        <polygon points="0 0, 10 3.5, 0 7"></polygon>
+        <polygon points="0 0, 10 3.5, 0 7" />
       </marker>
     </defs>
   );
