+++ src/types/Point.ts
@@ -1,1 +1,4 @@
-export type Point = { x: number; y: number };
+export interface Point {
+  x: number;
+  y: number;
+}
